+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased][unreleased]
+
+**Added**
+
+* Show basic move interpretation when no comment is provided
+
+**Changed**
+
+* Speed up SGF parsing
+* Speed up scoring
+* Show passes as squares in game graph
+
+**Fixed**
+
+* Fix regression when passing results in an uncatched exception
+
 ## [Sabaki v0.9.0][v0.9.0] (2016-02-20)
 
 **Added**
@@ -301,7 +317,7 @@ All notable changes to this project will be documented in this file.
 First release
 
 [unreleased]: https://github.com/yishn/Sabaki/compare/v0.9.0...master
-[v0.8.1]: https://github.com/yishn/Sabaki/compare/v0.8.1...v0.9.0
+[v0.9.0]: https://github.com/yishn/Sabaki/compare/v0.8.1...v0.9.0
 [v0.8.1]: https://github.com/yishn/Sabaki/compare/v0.8.0...v0.8.1
 [v0.8.0]: https://github.com/yishn/Sabaki/compare/v0.7.7...v0.8.0
 [v0.7.7]: https://github.com/yishn/Sabaki/compare/v0.7.6...v0.7.7